Fix QR code flow inside onboarding
